refactor: use standard folder structure and remove all eslint-disable any
