+++ src/app/services/http-stalker.service.ts
@@ -1,7 +1,6 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {User} from '../classes/user';
+import {User} from '../models/user';
 @Injectable({
   providedIn: 'root',
 })
