+++ src/app/services/login.service.ts
@@ -1,10 +1,9 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import {Injectable} from '@angular/core';
 import {catchError} from 'rxjs/operators';
 import {Observable, of} from 'rxjs';
-import {User} from '../classes/user';
+import {User} from '../models/user';
 import {environment} from '../../environments/environment';
-import {HttpStalkerService} from '../http-stalker.service/http-stalker.service';
+import {HttpStalkerService} from './http-stalker.service';
 
 @Injectable({
   providedIn: 'root',
